+++ App.js
@@ -1,6 +1,7 @@
 import { StyleSheet } from "react-native";
 import React from "react";
 import Navigation from "./config/navigation/Navigation";
+import { initializeApp } from "firebase/app";
 import { app } from "./config/utils/Firebase";
 
 export default function App() {
