+++ package.json
@@ -21,6 +21,7 @@
     "lodash": "^4.17.21",
     "react": "18.1.0",
     "react-native": "0.70.5",
+    "react-native-async-storage": "^0.0.1",
     "react-native-elements": "^3.4.3",
     "react-native-keyboard-aware-scroll-view": "^0.9.5",
     "react-native-safe-area-context": "4.4.1",
